+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue #210: Motor controller configuration
 - Issue #216: Change motors configuration for acceleration limits
+- Issue #218: Add move command monitoring
 
 ### Fixed